2025.12.19
1. TB_FLIGHT_OFFER 테이블, 시퀀스, 외래키 create
2. 검색 트리거시 api호출해서 해당 테이블로 필터조건에 해당하는 데이터 insert

+++ BackEnd_TripType_Final_Project/src/main/java/com/kh/triptype/admin/pricing/model/vo/FlightSearchHistoryVo.java
@@ -36,7 +36,7 @@ public class FlightSearchHistoryVo {
     private Date searchLogDate;
 
     /** 회원 번호 */
-    private int memberNo;
+    private Long memberNo;
 
     /** 출발 공항 IATA 코드 */
     private String departIataCode;
